cloud codestyle
