+++ cloud-minecraft/cloud-minestom/src/main/java/cloud/commandframework/minestom/MinestomCloudCommand.java
@@ -1,13 +1,49 @@
+//
+// MIT License
+//
+// Copyright (c) 2022 Alexander Söderberg & Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
 package cloud.commandframework.minestom;
 
-import cloud.commandframework.minestom.utils.ExceptionUtils;
+import cloud.commandframework.exceptions.*;
+import cloud.commandframework.execution.CommandResult;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.format.NamedTextColor;
 import net.minestom.server.command.CommandSender;
 import net.minestom.server.command.builder.Command;
 import net.minestom.server.command.builder.CommandContext;
+import org.checkerframework.checker.nullness.qual.NonNull;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import java.util.concurrent.CompletionException;
+import java.util.function.BiConsumer;
 
 public class MinestomCloudCommand<C> extends Command {
+
+    private static final String MESSAGE_INTERNAL_ERROR = "An internal error occurred while attempting to perform this command.";
+    private static final String MESSAGE_NO_PERMS =
+            "I'm sorry, but you do not have permission to perform this command. "
+                    + "Please contact the server administrators if you believe that this is in error.";
+    private static final String MESSAGE_UNKNOWN_COMMAND = "Unknown command. Type \"/help\" for help.";
+
     private final MinestomCommandManager<C> manager;
 
     public MinestomCloudCommand(MinestomCommandManager<C> manager, @NotNull String name, @Nullable String... aliases) {
@@ -18,8 +54,107 @@ public class MinestomCloudCommand<C> extends Command {
     @Override
     public void globalListener(@NotNull CommandSender commandSender, @NotNull CommandContext context, @NotNull String command) {
         C sender = manager.mapCommandSender(commandSender);
-        manager.executeCommand(sender, command).whenComplete((c, t) -> {
-            if (t != null) ExceptionUtils.onException(t, manager, sender, commandSender);
-        });
+        manager.executeCommand(sender, command).whenComplete(getResultConsumer(commandSender, sender));
+    }
+
+    private @NonNull BiConsumer<@NonNull CommandResult<C>, ? super Throwable> getResultConsumer(
+            final @NonNull CommandSender commandSender,
+            final @NonNull C sender
+    ) {
+        return (result, throwable) -> {
+            if (throwable != null) {
+                if (throwable instanceof CompletionException) {
+                    throwable = throwable.getCause();
+                }
+                final Throwable finalThrowable = throwable;
+                if (throwable instanceof InvalidSyntaxException) {
+                    this.manager.handleException(
+                            sender,
+                            InvalidSyntaxException.class,
+                            (InvalidSyntaxException) throwable,
+                            (c, e) ->
+                                    commandSender.sendMessage(
+                                            Component
+                                                    .text()
+                                                    .append(
+                                                            Component.text(
+                                                                    "Invalid Command Syntax. Correct command syntax is: ",
+                                                                    NamedTextColor.RED
+                                                            )
+                                                    ).append(
+                                                            Component.text(
+                                                                    e.getCorrectSyntax(),
+                                                                    NamedTextColor.GRAY
+                                                            )
+                                                    ).build()
+                                    )
+                    );
+                } else if (throwable instanceof InvalidCommandSenderException) {
+                    this.manager.handleException(
+                            sender,
+                            InvalidCommandSenderException.class,
+                            (InvalidCommandSenderException) throwable,
+                            (c, e) ->
+                                    commandSender.sendMessage(
+                                            Component.text(
+                                                    finalThrowable.getMessage(),
+                                                    NamedTextColor.RED
+                                            )
+                                    )
+                    );
+                } else if (throwable instanceof NoPermissionException) {
+                    this.manager.handleException(
+                            sender,
+                            NoPermissionException.class,
+                            (NoPermissionException) throwable,
+                            (c, e) -> commandSender.sendMessage(Component.text(MESSAGE_NO_PERMS))
+                    );
+                } else if (throwable instanceof NoSuchCommandException) {
+                    this.manager.handleException(
+                            sender,
+                            NoSuchCommandException.class,
+                            (NoSuchCommandException) throwable,
+                            (c, e) -> commandSender.sendMessage(Component.text(MESSAGE_UNKNOWN_COMMAND))
+                    );
+                } else if (throwable instanceof ArgumentParseException) {
+                    this.manager.handleException(
+                            sender,
+                            ArgumentParseException.class,
+                            (ArgumentParseException) throwable,
+                            (c, e) -> commandSender.sendMessage(
+                                    Component.text()
+                                            .append(Component.text(
+                                                    "Invalid Command Argument: ",
+                                                    NamedTextColor.RED
+                                            ))
+                                            .append(Component.text(
+                                                    finalThrowable.getCause().getMessage(),
+                                                    NamedTextColor.GRAY
+                                            ))
+                            )
+                    );
+                } else if (throwable instanceof CommandExecutionException) {
+                    this.manager.handleException(
+                            sender,
+                            CommandExecutionException.class,
+                            (CommandExecutionException) throwable,
+                            (c, e) -> {
+                                commandSender.sendMessage(
+                                        Component.text(
+                                                MESSAGE_INTERNAL_ERROR,
+                                                NamedTextColor.RED
+                                        )
+                                );
+                                finalThrowable.getCause().printStackTrace();
+                            }
+                    );
+                } else {
+                    commandSender.sendMessage(
+                            Component.text(MESSAGE_INTERNAL_ERROR, NamedTextColor.RED)
+                    );
+                    throwable.printStackTrace();
+                }
+            }
+        };
     }
 }
