+++ cloud-minecraft/cloud-minestom/src/main/java/cloud/commandframework/minestom/arguments/PlayerArgument.java
@@ -1,0 +1,217 @@
+//
+// MIT License
+//
+// Copyright (c) 2022 Alexander Söderberg & Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+package cloud.commandframework.minestom.arguments;
+
+import cloud.commandframework.ArgumentDescription;
+import cloud.commandframework.arguments.CommandArgument;
+import cloud.commandframework.arguments.parser.ArgumentParseResult;
+import cloud.commandframework.arguments.parser.ArgumentParser;
+import cloud.commandframework.captions.CaptionVariable;
+import cloud.commandframework.context.CommandContext;
+import cloud.commandframework.exceptions.parsing.NoInputProvidedException;
+import cloud.commandframework.exceptions.parsing.ParserException;
+import cloud.commandframework.minestom.MinestomCaptionKeys;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Queue;
+import java.util.function.BiFunction;
+import net.minestom.server.MinecraftServer;
+import net.minestom.server.entity.Player;
+import org.apiguardian.api.API;
+import org.checkerframework.checker.nullness.qual.NonNull;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+/**
+ * Argument that parses into a {@link Player}
+ *
+ * @param <C> Command sender type
+ * @since 1.9.0
+ */
+@SuppressWarnings("unused")
+public final class PlayerArgument<C> extends CommandArgument<C, Player> {
+
+    private PlayerArgument(
+            final boolean required,
+            final @NonNull String name,
+            final @NonNull String defaultValue,
+            final @Nullable BiFunction<@NonNull CommandContext<C>, @NonNull String,
+                    @NonNull List<@NonNull String>> suggestionsProvider,
+            final @NonNull ArgumentDescription defaultDescription
+    ) {
+        super(required, name, new PlayerParser<>(), defaultValue, Player.class, suggestionsProvider, defaultDescription);
+    }
+
+    /**
+     * Create a new {@link Builder}.
+     *
+     * @param name argument name
+     * @param <C>  sender type
+     * @return new {@link Builder}
+     */
+    @API(status = API.Status.STABLE)
+    public static <C> @NonNull Builder<C> builder(final @NonNull String name) {
+        return new Builder<>(name);
+    }
+
+    /**
+     * Create a new required command component
+     *
+     * @param name Component name
+     * @param <C>  Command sender type
+     * @return Created component
+     */
+    public static <C> @NonNull CommandArgument<C, Player> of(final @NonNull String name) {
+        return PlayerArgument.<C>builder(name).asRequired().build();
+    }
+
+    /**
+     * Create a new optional command component
+     *
+     * @param name Component name
+     * @param <C>  Command sender type
+     * @return Created component
+     */
+    public static <C> @NonNull CommandArgument<C, Player> optional(final @NonNull String name) {
+        return PlayerArgument.<C>builder(name).asOptional().build();
+    }
+
+    /**
+     * Create a new required command component with a default value
+     *
+     * @param name          Component name
+     * @param defaultPlayer Default player
+     * @param <C>           Command sender type
+     * @return Created component
+     */
+    public static <C> @NonNull CommandArgument<C, Player> optional(
+            final @NonNull String name,
+            final @NonNull String defaultPlayer
+    ) {
+        return PlayerArgument.<C>builder(name).asOptionalWithDefault(defaultPlayer).build();
+    }
+
+
+    public static final class Builder<C> extends CommandArgument.Builder<C, Player> {
+
+        private Builder(final @NonNull String name) {
+            super(Player.class, name);
+        }
+
+        /**
+         * Builder a new boolean component
+         *
+         * @return Constructed component
+         */
+        @Override
+        public @NonNull PlayerArgument<C> build() {
+            return new PlayerArgument<>(
+                    this.isRequired(),
+                    this.getName(),
+                    this.getDefaultValue(),
+                    this.getSuggestionsProvider(),
+                    this.getDefaultDescription()
+            );
+        }
+    }
+
+
+    public static final class PlayerParser<C> implements ArgumentParser<C, Player> {
+
+        @Override
+        public @NonNull ArgumentParseResult<Player> parse(
+                final @NonNull CommandContext<C> commandContext,
+                final @NonNull Queue<@NonNull String> inputQueue
+        ) {
+            final String input = inputQueue.peek();
+            if (input == null) {
+                return ArgumentParseResult.failure(new NoInputProvidedException(
+                        PlayerParser.class,
+                        commandContext
+                ));
+            }
+
+
+            Player player = MinecraftServer.getConnectionManager().getPlayer(input);
+
+            if (player == null) {
+                return ArgumentParseResult.failure(new PlayerParseException(input, commandContext));
+            }
+
+            inputQueue.remove();
+            return ArgumentParseResult.success(player);
+        }
+
+        @Override
+        public @NonNull List<@NonNull String> suggestions(
+                final @NonNull CommandContext<C> commandContext,
+                final @NonNull String input
+        ) {
+            List<String> output = new ArrayList<>();
+
+            for (Player player : MinecraftServer.getConnectionManager().getOnlinePlayers()) {
+                output.add(player.getUsername());
+            }
+
+            return output;
+        }
+    }
+
+
+    /**
+     * Player parse exception
+     */
+    public static final class PlayerParseException extends ParserException {
+
+        private static final long serialVersionUID = 2044681000582394398L;
+        private final String input;
+
+        /**
+         * Construct a new Player parse exception
+         *
+         * @param input   String input
+         * @param context Command context
+         */
+        public PlayerParseException(
+                final @NonNull String input,
+                final @NonNull CommandContext<?> context
+        ) {
+            super(
+                    PlayerParser.class,
+                    context,
+                    MinestomCaptionKeys.ARGUMENT_PARSE_FAILURE_PLAYER,
+                    CaptionVariable.of("input", input)
+            );
+            this.input = input;
+        }
+
+        /**
+         * Get the supplied input
+         *
+         * @return String value
+         */
+        public @NonNull String getInput() {
+            return this.input;
+        }
+    }
+}
