+++ cloud-minecraft/cloud-minestom/src/main/java/cloud/commandframework/minestom/arguments/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Minestom specific argument types
+ */
+package cloud.commandframework.minestom.arguments;
