--- cloud-minecraft/cloud-minestom/src/main/java/cloud/commandframework/minestom/caption/MinestomCaptionKeys.java
@@ -1,12 +1,0 @@
-package cloud.commandframework.minestom.caption;
-
-import cloud.commandframework.captions.Caption;
-
-import static cloud.commandframework.captions.Caption.of;
-
-public class MinestomCaptionKeys {
-    /**
-     * Variables: {input}
-     */
-    public static final Caption ARGUMENT_PARSE_FAILURE_PLAYER = of("argument.parse.failure.player");
-}
