+++ cloud-minecraft/cloud-minestom/src/main/java/cloud/commandframework/minestom/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Minestom implementation of cloud
+ */
+package cloud.commandframework.minestom;
